requirements.txt
accelerate==1.9.0
deepspeed==0.17.4
einops==0.8.0
transformers==4.55.0
huggingface-hub==0.34.3
optimum-quanto==0.2.7
datasets==4.0.0
omegaconf==2.3.0
sentencepiece==0.2.0
opencv-python==4.12.0.88
matplotlib==3.10.5
onnxruntime==1.22.1
torchvision==0.22.1
peft==0.17.0
timm==1.0.19
qwen_vl_utils==0.0.11
bitsandbytes
wandb
diffusers @ git+https://github.com/huggingface/diffusers@ba2ba9019f76fd96c532240ed07d3f98343e4041
